+++ packages/effects/layouts/src/basic/header/header.vue
@@ -2,15 +2,20 @@
 import { computed, useSlots } from 'vue';
 
 import { useRefresh } from '@vben/hooks';
-import { RotateCw } from '@vben/icons';
+import { LockKeyhole, LogOut, RotateCw } from '@vben/icons';
+import { $t } from '@vben/locales';
 import { preferences, usePreferences } from '@vben/preferences';
 import { useAccessStore } from '@vben/stores';
 
+import { useVbenModal } from '@vben-core/popup-ui';
 import { VbenFullScreen, VbenIconButton } from '@vben-core/shadcn-ui';
+
+import { useMagicKeys, whenever } from '@vueuse/core';
 
 import {
   GlobalSearch,
   LanguageToggle,
+  LockScreenModal,
   PreferencesButton,
   ThemeToggle,
   TimezoneButton,
@@ -18,9 +23,17 @@ import {
 
 interface Props {
   /**
+   * 头像
+   */
+  avatar?: string;
+  /**
    * Logo 主题
    */
   theme?: string;
+  /**
+   * 用户文本
+   */
+  text?: string;
 }
 
 defineOptions({
@@ -28,17 +41,78 @@ defineOptions({
 });
 
 withDefaults(defineProps<Props>(), {
+  avatar: '',
   theme: 'light',
+  text: '',
 });
 
-const emit = defineEmits<{ clearPreferencesAndLogout: [] }>();
+const emit = defineEmits<{
+  clearPreferencesAndLogout: [];
+  logout: [];
+  openLockScreen: [];
+}>();
 
 const REFERENCE_VALUE = 100;
 
 const accessStore = useAccessStore();
-const { globalSearchShortcutKey, preferencesButtonPosition } = usePreferences();
+const {
+  globalLockScreenShortcutKey,
+  globalLogoutShortcutKey,
+  globalSearchShortcutKey,
+  preferencesButtonPosition,
+} = usePreferences();
 const slots = useSlots();
 const { refresh } = useRefresh();
+
+const [LockModal, lockModalApi] = useVbenModal({
+  connectedComponent: LockScreenModal,
+});
+const [LogoutModal, logoutModalApi] = useVbenModal({
+  onConfirm() {
+    handleSubmitLogout();
+  },
+});
+
+function handleOpenLock() {
+  lockModalApi.open();
+}
+
+function handleSubmitLock(lockScreenPassword: string) {
+  lockModalApi.close();
+  accessStore.lockScreen(lockScreenPassword);
+}
+
+function handleLogout() {
+  logoutModalApi.open();
+}
+
+function handleSubmitLogout() {
+  emit('logout');
+  logoutModalApi.close();
+}
+
+// 快捷键
+if (preferences.shortcutKeys.enable) {
+  const keys = useMagicKeys();
+  const lockKey = keys['Alt+KeyL'];
+  const logoutKey = keys['Alt+KeyQ'];
+
+  if (lockKey) {
+    whenever(lockKey, () => {
+      if (globalLockScreenShortcutKey?.value) {
+        handleOpenLock();
+      }
+    });
+  }
+
+  if (logoutKey) {
+    whenever(logoutKey, () => {
+      if (globalLogoutShortcutKey?.value) {
+        handleLogout();
+      }
+    });
+  }
+}
 
 /**
  * 插槽列表类型
@@ -94,6 +168,18 @@ const rightSlots = computed(() => {
       name: 'notification',
     });
   }
+  // 锁定屏幕
+  if (preferences.widget.lockScreen) {
+    list.push({
+      index: REFERENCE_VALUE + 70,
+      name: 'lock-screen-btn',
+    });
+  }
+  // 退出登录
+  list.push({
+    index: REFERENCE_VALUE + 80,
+    name: 'logout-btn',
+  });
 
   Object.keys(slots).forEach((key) => {
     // 适配插槽名称，例如第一个插槽名：header-right-1
@@ -150,6 +236,26 @@ function clearPreferencesAndLogout() {
 </script>
 
 <template>
+  <LockModal
+    v-if="preferences.widget.lockScreen"
+    :avatar="avatar"
+    :text="text"
+    @submit="handleSubmitLock"
+  />
+
+  <LogoutModal
+    :cancel-text="$t('common.cancel')"
+    :confirm-text="$t('common.confirm')"
+    :fullscreen-button="false"
+    :title="$t('common.prompt')"
+    centered
+    content-class="px-8 min-h-10"
+    footer-class="border-none mb-3 mr-3"
+    header-class="border-none"
+  >
+    {{ $t('ui.widgets.logoutTip') }}
+  </LogoutModal>
+
   <template
     v-for="slot in leftSlots.filter((item) => item.index < REFERENCE_VALUE)"
     :key="slot.name"
@@ -205,6 +311,24 @@ function clearPreferencesAndLogout() {
         </template>
         <template v-else-if="slot.name === 'timezone'">
           <TimezoneButton class="mt-0.5 mr-1" />
+        </template>
+        <template v-else-if="slot.name === 'lock-screen-btn'">
+          <VbenIconButton
+            class="mr-1"
+            :tooltip="$t('ui.widgets.lockScreen.title')"
+            @click="handleOpenLock"
+          >
+            <LockKeyhole class="size-4" />
+          </VbenIconButton>
+        </template>
+        <template v-else-if="slot.name === 'logout-btn'">
+          <VbenIconButton
+            class="mr-1"
+            :tooltip="$t('common.logout')"
+            @click="handleLogout"
+          >
+            <LogOut class="size-4" />
+          </VbenIconButton>
         </template>
       </slot>
     </template>
